QC-15 QC-13 Update the booking and payment route to only save the booking at the end
