+++ react/qa-cinema/src/components/Navbar.jsx
@@ -24,7 +24,7 @@ const Navbar = () => {
                                 <Link className="nav-link" to="/listings/upcoming">Upcoming</Link>
                             </li>
                             <li className="nav-item">
-                                <Link className="nav-link" to="/bookings/makebooking">Booking</Link>
+                                <Link className="nav-link" to="/makebooking">Booking</Link>
                             </li>
                             <li className="nav-item">
                                 <Link className="nav-link" to="/discussion">Discussions</Link>
